# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"errors": {},
					"path": {
						"type": "string"
					},
					"timestamp": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"path",
					"timestamp"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"BillingUnitEnum": {
				"enum": [
					"day",
					"month",
					"one_time"
				],
				"type": "string"
			},
			"FacilityUnitTypeOfferingStatusEnum": {
				"enum": [
					"active",
					"inactive"
				],
				"type": "string"
			},
...
			"FacilityUnitTypeOfferingResponse": {
				"$ref": "#/components/schemas/infer_typeoffacilityUnitTypeOfferingResponseSchema_"
			},
			"PaginatedData_FacilityUnitTypeOfferingResponse_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/FacilityUnitTypeOfferingResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"items",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PaginatedData_FacilityUnitTypeOfferingResponse__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PaginatedData_FacilityUnitTypeOfferingResponse_"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_FacilityUnitTypeOfferingResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/FacilityUnitTypeOfferingResponse"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"infer_typeofcreateOfferingSchema_": {
				"properties": {
					"notes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/FacilityUnitTypeOfferingStatusEnum"
					},
					"minRentalDays": {
						"type": "number",
						"format": "double"
					},
					"depositMultiplier": {
						"type": "number",
						"format": "double"
					},
					"pricePerUnit": {
						"type": "number",
						"format": "double"
					},
					"billingUnit": {
						"$ref": "#/components/schemas/BillingUnitEnum"
					},
					"unitTypeId": {
						"type": "string"
					},
					"facilityId": {
						"type": "string"
					}
				},
				"required": [
					"minRentalDays",
					"depositMultiplier",
					"pricePerUnit",
					"billingUnit",
					"unitTypeId",
					"facilityId"
				],
				"type": "object"
			},
			"CreateOfferingRequest": {
				"$ref": "#/components/schemas/infer_typeofcreateOfferingSchema_"
			},
			"infer_typeofupdateOfferingSchema_": {
				"properties": {
					"notes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/FacilityUnitTypeOfferingStatusEnum"
					},
					"minRentalDays": {
						"type": "number",
						"format": "double"
					},
					"depositMultiplier": {
						"type": "number",
						"format": "double"
					},
					"pricePerUnit": {
						"type": "number",
						"format": "double"
					},
					"billingUnit": {
						"$ref": "#/components/schemas/BillingUnitEnum"
					}
				},
				"type": "object"
			},
			"UpdateOfferingRequest": {
				"$ref": "#/components/schemas/infer_typeofupdateOfferingSchema_"
			},
			"FacilityAmenityOfferingStatusEnum": {
				"enum": [
					"active",
					"inactive",
					"out_of_stock"
				],
				"type": "string"
			},
...
			"FacilityAmenityOfferingResponse": {
				"$ref": "#/components/schemas/infer_typeoffacilityAmenityOfferingResponseSchema_"
			},
			"PaginatedData_FacilityAmenityOfferingResponse_": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/FacilityAmenityOfferingResponse"
						},
						"type": "array"
					},
					"pagination": {
						"$ref": "#/components/schemas/PaginationMeta"
					}
				},
				"required": [
					"items",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_PaginatedData_FacilityAmenityOfferingResponse__": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/PaginatedData_FacilityAmenityOfferingResponse_"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse_FacilityAmenityOfferingResponse-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/FacilityAmenityOfferingResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_FacilityAmenityOfferingResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					},
					"messageCode": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/FacilityAmenityOfferingResponse"
					}
				},
				"required": [
					"success",
					"statusCode",
					"messageCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"infer_typeofcreateFacilityAmenityOfferingSchema_": {
				"properties": {
					"notes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/FacilityAmenityOfferingStatusEnum"
					},
					"totalQuantity": {
						"type": "number",
						"format": "double"
					},
					"billingUnit": {
						"$ref": "#/components/schemas/BillingUnitEnum"
					},
					"pricePerUnit": {
						"type": "number",
						"format": "double"
					},
					"amenityId": {
						"type": "string"
					},
					"facilityId": {
						"type": "string"
					}
				},
				"required": [
					"status",
					"totalQuantity",
					"billingUnit",
					"pricePerUnit",
					"amenityId",
					"facilityId"
				],
				"type": "object"
			},
			"CreateFacilityAmenityOfferingRequest": {
				"$ref": "#/components/schemas/infer_typeofcreateFacilityAmenityOfferingSchema_"
			},
			"infer_typeofupdateFacilityAmenityOfferingSchema_": {
				"properties": {
					"notes": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/FacilityAmenityOfferingStatusEnum"
					},
					"totalQuantity": {
						"type": "number",
						"format": "double"
					},
					"billingUnit": {
						"$ref": "#/components/schemas/BillingUnitEnum"
					},
					"pricePerUnit": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object"
			},
			"UpdateFacilityAmenityOfferingRequest": {
				"$ref": "#/components/schemas/infer_typeofupdateFacilityAmenityOfferingSchema_"
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter your JWT access token"
			},
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "accessToken",
				"description": "Access token stored in HttpOnly cookie"
			}
		}
	},
	"info": {
		"title": "project_be",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Van Thai"
		}
	},
	"paths": {
...
		"/facility-unit-type-offerings": {
			"get": {
				"operationId": "GetOfferings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedData_FacilityUnitTypeOfferingResponse__"
								}
							}
						}
					},
					"400": {
						"description": "Invalid query parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get list of facility unit type offerings (with pagination, filter, sorting)",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/FacilityUnitTypeOfferingStatusEnum"
						}
					},
					{
						"in": "query",
						"name": "billingUnit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BillingUnitEnum"
						}
					},
					{
						"in": "query",
						"name": "unitTypeId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "facilityId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"createdAt",
								"updatedAt",
								"pricePerUnit",
								"minRentalDays"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOffering",
				"responses": {
					"201": {
						"description": "Offering created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error, inactive Facility/UnitType, or offering already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility or Unit Type not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Create a new facility unit type offering",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOfferingRequest"
							}
						}
					}
				}
			}
		},
		"/facility-unit-type-offerings/{id}": {
			"get": {
				"operationId": "GetOfferingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get facility unit type offering details by ID",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update facility unit type offering",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateOfferingRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"description": "Soft-delete facility unit type offering",
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/facility-unit-type-offerings/{id}/restore": {
			"post": {
				"operationId": "RestoreOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityUnitTypeOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Offering is not deleted or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Unit Type Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-amenity-offerings": {
			"get": {
				"operationId": "GetOfferings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_PaginatedData_FacilityAmenityOfferingResponse__"
								}
							}
						}
					},
					"400": {
						"description": "Invalid query parameters",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "maxPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "minPrice",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/FacilityAmenityOfferingStatusEnum"
						}
					},
					{
						"in": "query",
						"name": "billingUnit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/BillingUnitEnum"
						}
					},
					{
						"in": "query",
						"name": "amenityId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "facilityId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"status",
								"createdAt",
								"updatedAt",
								"pricePerUnit",
								"totalQuantity",
								"inUseQuantity"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateOffering",
				"responses": {
					"201": {
						"description": "Facility amenity offering created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or offering already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateFacilityAmenityOfferingRequest"
							}
						}
					}
				}
			}
		},
		"/facility-amenity-offerings/facility/{facilityId}/available": {
			"get": {
				"operationId": "GetAvailableOfferings",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse-Array_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid Facility ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "facilityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-amenity-offerings/facility/{facilityId}/amenity/{amenityId}": {
			"get": {
				"operationId": "GetOfferingByFacilityAndAmenity",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid parameters format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "facilityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "amenityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-amenity-offerings/{id}": {
			"get": {
				"operationId": "GetOfferingById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Validation error or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateFacilityAmenityOfferingRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid ID format or items in use",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/facility-amenity-offerings/{id}/restore": {
			"post": {
				"operationId": "RestoreOffering",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_FacilityAmenityOfferingResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Offering is not deleted or invalid ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "Facility Amenity Offering not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Facility Amenity Offerings"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "https://[object Object]/"
		}
	]
}